requirements.txt
fastapi
uvicorn
python-dotenv
pandas
openpyxl
ollama